+++ packages/backend/src/nest/storage/certifacteRequests/certificatesRequestsStore.spec.ts
@@ -1,45 +1,49 @@
-import OrbitDB from 'orbit-db'
 import fs from 'fs'
 import { jest } from '@jest/globals'
 import { create, IPFS } from 'ipfs-core'
 import { CertificatesRequestsStore } from './certificatesRequestsStore'
 import { EventEmitter } from 'events'
-import { StorageEvents } from './storage.types'
-import { TestConfig } from '../const'
+import { StorageEvents } from '../storage.types'
+import { TestConfig } from '../../const'
+import { Test, TestingModule } from '@nestjs/testing'
+import { TestModule } from '../../common/test.module'
+import { StorageModule } from '../storage.module'
+import { OrbitDb } from '../orbitDb/orbitDb.service'
+import PeerId from 'peer-id'
 
-// Futher improvement is creating 2 orbitdb instances, connect them and create 'real-life' tests.
-const createOrbitDbInstance = async () => {
-  const ipfs: IPFS = await create()
-  // @ts-ignore
-  const orbitdb = await OrbitDB.createInstance(ipfs, {
-    directory: TestConfig.ORBIT_DB_DIR,
-  })
-
-  return { orbitdb, ipfs }
-}
-
-const replicatedEvent = async (store: any) => {
-  store.events.emit('replicated')
+const replicatedEvent = async (certificatesRequestsStore: any) => {
+  certificatesRequestsStore.events.emit('replicated')
   await new Promise<void>(resolve => setTimeout(() => resolve(), 2000))
 }
 
 describe('CertificatesRequestsStore', () => {
+  let module: TestingModule
+  let certificatesRequestsStore: CertificatesRequestsStore
+  let orbitDb: OrbitDb
   let ipfs: IPFS
-  let orbitdb: OrbitDB
-  let store: CertificatesRequestsStore
-  let emitter: EventEmitter
 
   beforeEach(async () => {
-    ;({ orbitdb, ipfs } = await createOrbitDbInstance())
-    store = new CertificatesRequestsStore(orbitdb)
-    emitter = new EventEmitter()
-    await store.init(emitter)
+    jest.clearAllMocks()
+
+    module = await Test.createTestingModule({
+      imports: [TestModule, StorageModule],
+    }).compile()
+
+    certificatesRequestsStore = await module.resolve(CertificatesRequestsStore)
+
+    orbitDb = await module.resolve(OrbitDb)
+    const peerId = await PeerId.create()
+    ipfs = await create()
+    await orbitDb.create(peerId, ipfs)
+
+    const emitter = new EventEmitter()
+    await certificatesRequestsStore.init(emitter)
   })
 
   afterEach(async () => {
-    await store.close()
-    await orbitdb.stop()
+    await orbitDb.stop()
     await ipfs.stop()
+    await certificatesRequestsStore.close()
     if (fs.existsSync(TestConfig.ORBIT_DB_DIR)) {
       fs.rmSync(TestConfig.ORBIT_DB_DIR, { recursive: true })
     }
@@ -59,9 +63,8 @@ describe('CertificatesRequestsStore', () => {
   })
 
   it('getCsrs - remove old csrs and replace with new for each pubkey', async () => {
-    const store = new CertificatesRequestsStore(orbitdb)
     const emitter = new EventEmitter()
-    await store.init(emitter)
+    await certificatesRequestsStore.init(emitter)
 
     const allCsrs = [
       'MIIDITCCAsYCAQAwSTFHMEUGA1UEAxM+anR3c3hxMnZ1dWthY3JodWhvdnAzd2JxbzRxNXc0d2s3Nm1qbWJ3cXk3eGNma2FsdmRxb3hhYWQub25pb24wWTATBgcqhkjOPQIBBggqhkjOPQMBBwNCAAQE2q6iS+WCmIVCSFI2AjHrW6ujUdrceD5T2xkcTJBTn0y50WphcupUajCRgkXaTBkTsGNJ3qWRZAKX7CiuehBJoIICGTAuBgkqhkiG9w0BCQ4xITAfMB0GA1UdDgQWBBQuE5JgPY/BYBpgG5pnjMkEEIkrGjCCAUcGCSqGSIb3DQEJDDGCATgEggE0BDlx84glBl72q82F2a+y8iTVKM8IMiXYYrmNyhFPj6XsfVQpvLhNviZ5zHdMBWbFj44vTSUIasNP9I9eCWSEAaEJqjngEh18WCRS/XbvQxI/8qB5pzcfghvM8BCgSLbSEjK2GMYVhCXmRH1YGHIZu0+Ii9pe5nwG154JlPUsmIRgu6ruY6PQk65Aoo4OyhPn5CCUFInptHcz1JpAiCRe0Z6wuQHud03VY50fx4ETdmUNJBEIPOyd/Xn6lMOi6SaWGHbCWiufeJRm+mRdoHJAEt6kPLhGIYGyduNT/8cGoe2xKyQDvNoTr4dqqRZ2HgZ18nicsTHswpGqAlUnZXaA3V85Qu1cvaMAqEoPOUlGP9AriIVwtIZM0hdWHqKHgBCZrKfHb5oLxt6ourQ3+q19tvx+u6UwFAYKKwYBBAGDjBsCATEGEwRlbGxvMD0GCSsGAQIBDwMBATEwEy5RbVVvNXN0NXNqR3RFMUtQeXhOVW5pTWhnQXduV0JVNXk3TnpoMlpRRkdacVdiMEcGA1UdETFAEz5qdHdzeHEydnV1a2Fjcmh1aG92cDN3YnFvNHE1dzR3azc2bWptYndxeTd4Y2ZrYWx2ZHFveGFhZC5vbmlvbjAKBggqhkjOPQQDAgNJADBGAiEAt9udw7B7vnjJyWvkkRLb7DImFXwsrSxirqbmhIH+1rUCIQD86GWyfAE2d8gCNAn4h1t9B+mAx33ZdPLgFssHl1i3pA==',
@@ -76,15 +79,15 @@ describe('CertificatesRequestsStore', () => {
     ]
 
     for (const csr of allCsrs) {
-      await store.addUserCsr(csr)
+      await certificatesRequestsStore.addUserCsr(csr)
       // This should not be there, there's bug in orbitdb, it breaks if we add entries without artificial sleep, tho it's awaited.
       // https://github.com/TryQuiet/quiet/issues/2121
       await new Promise<void>(resolve => setTimeout(() => resolve(), 500))
     }
 
-    await store.getCsrs()
+    await certificatesRequestsStore.getCsrs()
 
-    const filteredCsrs = await store.getCsrs()
+    const filteredCsrs = await certificatesRequestsStore.getCsrs()
 
     expect(filteredCsrs.length).toEqual(allCsrs.length - 1)
     expect(filteredCsrs).toEqual([
@@ -98,74 +101,69 @@ describe('CertificatesRequestsStore', () => {
     ])
   })
   it('replicated event', async () => {
-    const store = new CertificatesRequestsStore(orbitdb)
     const emitter = new EventEmitter()
-    await store.init(emitter)
+    await certificatesRequestsStore.init(emitter)
 
     const spy = jest.fn()
 
     emitter.on(StorageEvents.LOADED_USER_CSRS, spy)
-    await replicatedEvent(store.store)
+    await replicatedEvent(certificatesRequestsStore.store)
 
     expect(spy).toBeCalledTimes(1)
   })
   it('2 replicated events - first not resolved ', async () => {
-    const store = new CertificatesRequestsStore(orbitdb)
     const emitter = new EventEmitter()
-    await store.init(emitter)
+    await certificatesRequestsStore.init(emitter)
 
     const spy = jest.fn()
 
     emitter.on(StorageEvents.LOADED_USER_CSRS, spy)
-    await replicatedEvent(store.store)
-    await replicatedEvent(store.store)
+    await replicatedEvent(certificatesRequestsStore.store)
+    await replicatedEvent(certificatesRequestsStore.store)
 
     expect(spy).toBeCalledTimes(1)
   })
   it('2 replicated events - first resolved ', async () => {
-    const store = new CertificatesRequestsStore(orbitdb)
     const emitter = new EventEmitter()
-    await store.init(emitter)
+    await certificatesRequestsStore.init(emitter)
 
     const spy = jest.fn()
 
     emitter.on(StorageEvents.LOADED_USER_CSRS, spy)
-    await replicatedEvent(store.store)
-    await replicatedEvent(store.store)
+    await replicatedEvent(certificatesRequestsStore.store)
+    await replicatedEvent(certificatesRequestsStore.store)
 
-    store.resolveCsrReplicatedPromise(1)
+    certificatesRequestsStore.resolveCsrReplicatedPromise(1)
     await new Promise<void>(resolve => setTimeout(() => resolve(), 500))
 
     expect(spy).toBeCalledTimes(2)
   })
   it('3 replicated events - no resolved promises', async () => {
-    const store = new CertificatesRequestsStore(orbitdb)
     const emitter = new EventEmitter()
-    await store.init(emitter)
+    await certificatesRequestsStore.init(emitter)
 
     const spy = jest.fn()
 
     emitter.on(StorageEvents.LOADED_USER_CSRS, spy)
-    await replicatedEvent(store.store)
-    await replicatedEvent(store.store)
-    await replicatedEvent(store.store)
+    await replicatedEvent(certificatesRequestsStore.store)
+    await replicatedEvent(certificatesRequestsStore.store)
+    await replicatedEvent(certificatesRequestsStore.store)
 
     expect(spy).toBeCalledTimes(1)
   })
   it('3 replicated events - two resolved promises ', async () => {
-    const store = new CertificatesRequestsStore(orbitdb)
     const emitter = new EventEmitter()
-    await store.init(emitter)
+    await certificatesRequestsStore.init(emitter)
 
     const spy = jest.fn()
     emitter.on(StorageEvents.LOADED_USER_CSRS, spy)
 
-    await replicatedEvent(store.store)
-    await replicatedEvent(store.store)
-    store.resolveCsrReplicatedPromise(1)
+    await replicatedEvent(certificatesRequestsStore.store)
+    await replicatedEvent(certificatesRequestsStore.store)
+    certificatesRequestsStore.resolveCsrReplicatedPromise(1)
     await new Promise<void>(resolve => setTimeout(() => resolve(), 500))
-    await replicatedEvent(store.store)
-    store.resolveCsrReplicatedPromise(2)
+    await replicatedEvent(certificatesRequestsStore.store)
+    certificatesRequestsStore.resolveCsrReplicatedPromise(2)
     await new Promise<void>(resolve => setTimeout(() => resolve(), 500))
 
     expect(spy).toBeCalledTimes(3)