+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Fix: joining with QR code after leaving community on mobile
 
+* Refactor: backend, storage module - extracting OrbitDB as another provider, refactor of  CertificatesRequestsStore, CommunityMetadataStore, CertificatesStore as Nest providers, store tests adjustments,  file structure
+
 [2.0.3-alpha.15]
 
 * Fix: construct all stores before initializing them - initializing community metadata store sets metadata in certificates store